# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sipelayar API",
    "version": "1.0.0",
    "description": "API documentation for Sipelayar application",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "V1 Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token issued on login"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key issued to clients"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/tokens": {
      "get": {
        "summary": "Get all tokens",
        "tags": [
          "Tokens"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "deviceId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by device ID"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "used",
                "unused"
              ]
            },
            "description": "Filter by token status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of tokens"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create a new token",
        "tags": [
          "Tokens"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "deviceId",
                  "amount"
                ],
                "properties": {
                  "deviceId": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number",
                    "format": "float",
                    "minimum": 0.01
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Device not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/tokens/{id}": {
      "get": {
        "summary": "Get token by ID",
        "tags": [
          "Tokens"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Token details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Token not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/tokens/device/{deviceId}": {
      "get": {
        "summary": "Get tokens for a specific device",
        "tags": [
          "Tokens"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "deviceId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Device ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "used",
                "unused"
              ]
            },
            "description": "Filter by token status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of tokens for the device"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Device not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/tokens/validate": {
      "post": {
        "summary": "Validate a token",
        "tags": [
          "Tokens"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "deviceKey"
                ],
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "deviceKey": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token validated successfully"
          },
          "400": {
            "description": "Invalid token or device key"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Device is inactive"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/tokens/stats": {
      "get": {
        "summary": "Get token statistics",
        "tags": [
          "Tokens"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token statistics"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}